p8

+++ config/puma.rb
@@ -40,7 +40,7 @@
 # Processes: ec2 default just deploys to 1 process engine- 
 
 workers Integer(8)
-threads_count = Integer(8)
+threads_count = Integer(1)
 threads threads_count, threads_count
 
 preload_app!
